move things around in internal/adminapi
